pyproject.toml
[project]
name = "summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "chromadb>=1.3.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.2",
    "groq>=0.33.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-yt-dlp>=0.0.8",
    "ollama>=0.6.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.3",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pytube>=15.0.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
    "whisper>=1.1.10",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.10.22",
]
